+++ src/components/web-server.ts
@@ -4,7 +4,7 @@ import { CustomSize } from '../types/size';
 import { PredefinedSize, commonTags } from '../constants';
 import { ContainerDefinition } from '@pulumi/aws/ecs';
 import { AcmCertificate } from './acm-certificate';
-import { EcsArgs, assumeRolePolicy, awsRegion } from '../common/ecs';
+import { EcsArgs, assumeRolePolicy, awsRegion } from './ecs';
 
 export type WebServerArgs = EcsArgs & {
   /**
